Add parameters for income and capital gains tax rates (#6)
Added the ability to set the income tax rate and capital gains tax rate as parameters, allowing users in different regions or with varying tax rates to generate reports with their values at the time of creation. Previously, this was only possible by manually changing the values in the generated Excel files.

The default values are set to the maximum tax rates in Germany.